refactor: complete Repository Pattern implementation for Bookcase and eliminate DTO leakage in Shelf
Context:
This refactor completes the hexagonal architecture alignment for both Bookcase
and Shelf modules by properly implementing the Repository Pattern and eliminating
architectural violations where DTOs leaked into the domain layer.

What changed:

**Domain Layer (core/domain):**
- Moved Bookcase entity to core/domain/model package for proper domain segregation
- Relocated BookcaseRepository interface to core/domain/ports/outbound
- Relocated BookAccessPort to core/ports (removing nested domain/ports structure)
- Updated ShelfDomainRepository to return domain Shelf objects instead of DTOs
  (findShelfSummariesByBookcaseId, getShelfShelfOptionResponse now return List<Shelf>)

**Application Layer (core/application):**
- Updated BookcaseService to reference domain port instead of infrastructure repository
- Enhanced ShelfService to perform DTO mapping at application boundary
  (getShelfSummariesForBookcase, getShelfOptionsByBookcase now map Shelf → DTO)

**Infrastructure Layer:**
- Updated BookcaseRepositoryImpl imports to reference relocated domain port
- Fixed ShelfDomainRepositoryImpl to return domain objects from repository methods
- Updated BookAccessAdapter to reference relocated BookAccessPort
- Fixed BookcaseMapper import for relocated Bookcase domain model

**Tests:**
- Updated BookcaseServiceTest and BookcaseTest imports for relocated domain classes
- Cleaned up verbose ShelfDomainRepository mock declarations in ShelfServiceTest

**Documentation:**
- Updated devlog with corrected Bookcase package path

Notes/Risks:
- Breaking change: ShelfDomainRepository contract changed (now returns domain objects)
- Adapters and services updated to handle new contract
- All tests pass with updated imports
- This establishes proper hexagonal boundaries: DTOs stay in API/application layer,
  domain objects stay in domain/application layer, entities stay in infrastructure
