+++ src/main/java/com/penrose/bibby/library/stacks/bookcase/core/application/BookcaseService.java
@@ -3,9 +3,9 @@ package com.penrose.bibby.library.stacks.bookcase.core.application;
 import com.penrose.bibby.library.stacks.bookcase.api.CreateBookcaseResult;
 import com.penrose.bibby.library.stacks.bookcase.api.dtos.BookcaseDTO;
 import com.penrose.bibby.library.stacks.bookcase.api.ports.inbound.BookcaseFacade;
+import com.penrose.bibby.library.stacks.bookcase.core.domain.ports.outbound.BookcaseRepository;
 import com.penrose.bibby.library.stacks.bookcase.infrastructure.entity.BookcaseEntity;
 import com.penrose.bibby.library.stacks.bookcase.infrastructure.repository.BookcaseJpaRepository;
-import com.penrose.bibby.library.stacks.bookcase.infrastructure.repository.BookcaseRepository;
 import com.penrose.bibby.library.stacks.shelf.api.ports.inbound.ShelfFacade;
 import java.util.List;
 import java.util.Optional;
