+++ src/main/java/com/penrose/bibby/library/stacks/bookcase/core/domain/ports/outbound/BookcaseRepository.java
@@ -1,4 +1,4 @@
-package com.penrose.bibby.library.stacks.bookcase.infrastructure.repository;
+package com.penrose.bibby.library.stacks.bookcase.core.domain.ports.outbound;
 
 import com.penrose.bibby.library.stacks.bookcase.infrastructure.entity.BookcaseEntity;
 import java.util.List;
