+++ src/main/java/com/penrose/bibby/library/stacks/bookcase/infrastructure/mapping/BookcaseMapper.java
@@ -1,6 +1,6 @@
 package com.penrose.bibby.library.stacks.bookcase.infrastructure.mapping;
 
-import com.penrose.bibby.library.stacks.bookcase.core.domain.Bookcase;
+import com.penrose.bibby.library.stacks.bookcase.core.domain.model.Bookcase;
 
 public class BookcaseMapper {
